fix(*): 优化代码
- 删除无效 throw
- Controller 中的 prefix 替换为 static final
- 删除多余的初始值
- 删除 <> 中的类名
- 没必要的包装类声明改成基础类型
- append(……) + "……" 的方式替换为链式调用，循环中的 += 替换为 append
- 删除多余的局部变量
- 删除接口和枚举中变量和方法多余的修饰符
- 给局部变量加 final
- 修改注释为 javadoc 形式
- 合并 catch 中的异常类
- 获取对象属性的回调替换为 lambda 或方法引用
- 循环 add 替换为 addAll
- 删除冗余装箱
- indexOf 替换为 contains
- StringBuffer 替换为 StringBuilder
- 简化 == true/false 这种判断……
- fori 替换为 foreach

+++ ruoyi-system/src/main/java/com/ruoyi/system/service/impl/SysConfigServiceImpl.java
@@ -144,9 +144,9 @@ public class SysConfigServiceImpl implements ISysConfigService {
    */
   @Override
   public String checkConfigKeyUnique(SysConfig config) {
-    Long configId = StringUtils.isNull(config.getConfigId()) ? -1L : config.getConfigId();
+    long configId = StringUtils.isNull(config.getConfigId()) ? -1L : config.getConfigId();
     SysConfig info = configMapper.checkConfigKeyUnique(config.getConfigKey());
-    if (StringUtils.isNotNull(info) && info.getConfigId().longValue() != configId.longValue()) {
+    if (StringUtils.isNotNull(info) && info.getConfigId() != configId) {
       return UserConstants.CONFIG_KEY_NOT_UNIQUE;
     }
     return UserConstants.CONFIG_KEY_UNIQUE;
